add user permission check when deleting an item

+++ src/resolvers/Mutation.js
@@ -41,8 +41,13 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. Find item
-    const item = await ctx.db.query.item({ where }, '{ id title }');
-    // 2. TODO: check if they own that item or have the permissions
+    const item = await ctx.db.query.item({ where }, '{ id title user { id } }');
+    // 2. Check if they own that item or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+    if (!ownsItem || !hasPermissions) {
+      throw new Error("You don't have permission to delete this item");
+    }
     // 3. Delete it
     return ctx.db.mutation.deleteItem({ where }, info);;
   },
